refactor: 계산 기능 Caculator -> OperatorType으로 변경
